+++ src/OrderLinesList/OrderLinesFiltersContainer.js
@@ -1,11 +1,15 @@
-import React from 'react';
-import PropTypes from 'prop-types';
 import { get } from 'lodash';
+import PropTypes from 'prop-types';
 
-import { stripesConnect } from '@folio/stripes/core';
+import {
+  checkIfUserInCentralTenant,
+  stripesConnect,
+  useStripes,
+} from '@folio/stripes/core';
 import {
   DICT_FUNDS,
   fundsManifest,
+  useCentralOrderingSettings,
 } from '@folio/stripes-acq-components';
 import { OrderLinesFilters } from '@folio/plugin-find-po-line';
 
@@ -14,8 +18,14 @@ import {
 } from '../components/Utils/resources';
 
 const OrderLinesFiltersContainer = ({ resources, activeFilters, applyFilters, disabled }) => {
+  const stripes = useStripes();
+
   const funds = get(resources, `${DICT_FUNDS}.records`);
   const materialTypes = get(resources, 'materialTypes.records', []);
+
+  const { enabled: isCentralOrderingEnabled } = useCentralOrderingSettings({
+    enabled: checkIfUserInCentralTenant(stripes),
+  });
 
   return (
     <OrderLinesFilters
@@ -24,6 +34,7 @@ const OrderLinesFiltersContainer = ({ resources, activeFilters, applyFilters, di
       activeFilters={activeFilters}
       applyFilters={applyFilters}
       disabled={disabled}
+      crossTenant={isCentralOrderingEnabled}
     />
   );
 };